@@ -1,5 +1,7 @@
 <?php
 
+namespace LaravelEnso\Core\Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -7,7 +9,7 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call([
-            RoleSeeder::class,
+            RoleSeeder::class, //TODO import
             UserGroupSeeder::class,
             UserSeeder::class,
             LanguageSeeder::class,
