@@ -1,5 +1,7 @@
 <?php
 
+namespace LaravelEnso\Core\Database\Seeders;
+
 use Illuminate\Database\Seeder;
 use LaravelEnso\Core\Models\UserGroup;
 use LaravelEnso\Roles\Models\Role;
@@ -8,7 +10,7 @@ class UserGroupSeeder extends Seeder
 {
     public function run()
     {
-        factory(UserGroup::class)->create([
+       UserGroup:: factory()->create([
             'name' => 'Administrators',
             'description' => 'Administrator users group',
         ])->roles()->sync(Role::pluck('id'));
